+++ R/raster-fns.R
@@ -3,7 +3,7 @@
 #' This function returns the `rasterFunctionInfos` field of the `ImageServer`'s metadata
 #' as a `data.frame`. If the field does not exist then an error is emitted.
 #'
-#' @inheritParams arcgisutils::infer_esri_type
+#' @inheritParams arcgisutils::as_fields
 #' @param x an `ImageServer`.
 #' @returns a data.frame of the available raster functions.
 #' @export
@@ -19,7 +19,8 @@
 #' raster_fns <- list_service_raster_fns(service)
 #' head(raster_fns)
 #' }
-list_service_raster_fns <- function(
+#' @rdname raster_fns
+list_raster_fns <- function(
   x,
   arg = rlang::caller_arg(x),
   call = rlang::caller_call()
@@ -30,4 +31,19 @@ list_service_raster_fns <- function(
     cli::cli_abort("{.arg arg} does not support raster functions")
   }
   data_frame(x$rasterFunctionInfos)
+}
+
+#' @export
+#' @rdname raster_fns
+list_service_raster_fns <- function(
+  x,
+  arg = rlang::caller_arg(x),
+  call = rlang::caller_call()
+) {
+  lifecycle::deprecate_soft(
+    "0.4.0",
+    "list_service_raster_fns()",
+    "list_raster_fns()"
+  )
+  list_raster_fns(x, arg, call)
 }