+++ NEWS.md
@@ -7,6 +7,7 @@
 - `update_features()` is now parallelized and sends updates in chunks see `chunk_size` argument and `progress` arguments. The return type is now a `data.frame` and not a list with `updateResults`
 - `delete_features()` is now parallelized and deletes in chunks. See above. 
 - `{arcgislayers}` now depends on R 4.2 or higher.
+- `list_service_raster_fns()` is now deprecated in favor of `list_raster_fns()`
 
 ## New features